refactor: rename app to meins2
