+++ meins-rn-ts/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.typescriptreactnativestarter",
+    package = "com.meins2",
 )
 
 android_resource(
     name = "res",
-    package = "com.typescriptreactnativestarter",
+    package = "com.meins2",
     res = "src/main/res",
 )
