+++ meins-rn-ts/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">TypescriptReactNativeStarter</string>
+    <string name="app_name">meins2</string>
 </resources>
