+++ meins-rn-ts/android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'TypescriptReactNativeStarter'
+rootProject.name = 'meins2'
 include ':react-native-vector-icons'
 project(':react-native-vector-icons').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-vector-icons/android')
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
