+++ meins-rn-ts/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "TypescriptReactNativeStarter",
-  "displayName": "TypescriptReactNativeStarter"
+  "name": "meins2",
+  "displayName": "meins2"
 }
