+++ meins-rn-ts/ios/meins2.xcodeproj/xcshareddata/xcschemes/meins2.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "TypescriptReactNativeStarter.app"
-               BlueprintName = "TypescriptReactNativeStarter"
-               ReferencedContainer = "container:TypescriptReactNativeStarter.xcodeproj">
+               BuildableName = "meins2.app"
+               BlueprintName = "meins2"
+               ReferencedContainer = "container:meins2.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "TypescriptReactNativeStarterTests.xctest"
-               BlueprintName = "TypescriptReactNativeStarterTests"
-               ReferencedContainer = "container:TypescriptReactNativeStarter.xcodeproj">
+               BuildableName = "meins2Tests.xctest"
+               BlueprintName = "meins2Tests"
+               ReferencedContainer = "container:meins2.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "TypescriptReactNativeStarter.app"
-            BlueprintName = "TypescriptReactNativeStarter"
-            ReferencedContainer = "container:TypescriptReactNativeStarter.xcodeproj">
+            BuildableName = "meins2.app"
+            BlueprintName = "meins2"
+            ReferencedContainer = "container:meins2.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "TypescriptReactNativeStarter.app"
-            BlueprintName = "TypescriptReactNativeStarter"
-            ReferencedContainer = "container:TypescriptReactNativeStarter.xcodeproj">
+            BuildableName = "meins2.app"
+            BlueprintName = "meins2"
+            ReferencedContainer = "container:meins2.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
