+++ meins-rn-ts/ios/meins2.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:TypescriptReactNativeStarter.xcodeproj">
+      location = "group:meins2.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
