+++ meins-rn-ts/ios/meins2.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
