+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/Contents.json
