+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/ios-marketing-1024x1024-1x.png
