+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/ipad-20x20-1x.png
