+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/ipad-29x29-2x.png
