+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/ipad-40x40-2x.png
