+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/ipad-50x50-1x.png
