+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/ipad-72x72-1x.png
