+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/ipad-76x76-1x.png
