+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/ipad-83.5x83.5-2x.png
