+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/iphone-20x20-3x.png
