+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/iphone-29x29-1x.png
