+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/iphone-40x40-2x.png
