+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/iphone-57x57-1x.png
