+++ meins-rn-ts/ios/meins2/Images.xcassets/AppIcon.appiconset/iphone-60x60-3x.png
