+++ meins-rn-ts/ios/meins2/Images.xcassets/BootSplashLogo.imageset/Contents.json
