+++ meins-rn-ts/ios/meins2/Images.xcassets/BootSplashLogo.imageset/bootsplash_logo.png
