+++ meins-rn-ts/ios/meins2/Images.xcassets/Contents.json
