+++ meins-rn-ts/ios/meins2/LaunchScreen.storyboard
