+++ meins-rn-ts/ios/meins2Tests/Info.plist
